+++ app/legacy/v2/notifications/rest.py
@@ -1,9 +1,7 @@
 """All endpoints for the v2/notifications route."""
 
-import asyncio
-from abc import ABC, abstractmethod
-from typing import Annotated, TypedDict
-from uuid import UUID, uuid4
+from typing import Annotated
+from uuid import uuid4
 
 from fastapi import APIRouter, BackgroundTasks, Body, Depends, Request, status
 from pydantic import UUID4
@@ -11,6 +9,11 @@ from starlette_context import context
 
 from app.auth import JWTBearer
 from app.constants import NotificationType
+from app.legacy.v2.notifications.handlers import (
+    DirectSmsNotificationHandler,
+    IdentifierSmsNotificationHandler,
+    SmsNotificationHandler,
+)
 from app.legacy.v2.notifications.route_schema import (
     HttpsUrl,
     V2PostPushRequestModel,
@@ -79,81 +82,6 @@ async def create_push_notification(
         template_id,
     )
     return V2PostPushResponseModel()
-
-
-class NotificationRecord(TypedDict, total=False):
-    """Type definition for notification records."""
-
-    id: str
-    template_id: str
-    template_version: str
-    recipient_identifier_type: str
-    recipient_identifier_value: str
-    recipient: str  # Add this field for direct SMS notifications
-    status: str
-    timestamp: str
-    reason: str
-    phone_number: str
-
-
-class SmsNotificationHandler(ABC):
-    """Abstract base class for handling SMS notifications."""
-
-    @abstractmethod
-    async def process(self, notification_id: UUID) -> None:
-        """Process the notification request.
-
-        Args:
-            notification_id (UUID): The generated notification ID
-        """
-        pass
-
-
-class DirectSmsNotificationHandler(SmsNotificationHandler):
-    """Handler for direct SMS notifications via phone number."""
-
-    def __init__(self, phone_number: str) -> None:
-        """Initialize with recipient phone number.
-
-        Args:
-            phone_number (str): The recipient's phone number
-        """
-        self.phone_number = phone_number
-
-    async def process(self, notification_id: UUID) -> None:
-        """Process a direct SMS notification.
-
-        Args:
-            notification_id (UUID): Generated notification ID
-        """
-        logger.info('Calling celery task deliver_sms with notification id {}', notification_id)
-
-        # Simulate an async operation
-        await asyncio.sleep(0.01)
-
-
-class IdentifierSmsNotificationHandler(SmsNotificationHandler):
-    """Handler for SMS notifications via recipient identifier."""
-
-    def __init__(self, recipient_identifier: dict) -> None:
-        """Initialize with recipient identifier.
-
-        Args:
-            recipient_identifier (dict): The recipient identifier dictionary
-        """
-        self.recipient_identifier = recipient_identifier
-
-    async def process(self, notification_id: UUID) -> None:
-        """Process an SMS notification via recipient identifier.
-
-        Args:
-            notification_id (UUID): Generated notification ID
-        """
-        logger.info('Calling celery task lookup_va_profile_id with notification id {}.', notification_id)
-        await asyncio.sleep(0.01)
-
-        logger.info('Calling celery task deliver_sms with notification id {}', notification_id)
-        await asyncio.sleep(0.01)
 
 
 def get_sms_notification_handler(request: V2PostSmsRequestModel) -> SmsNotificationHandler:
